# Cache/points counting refactoring + related bug fixes