--- a/course/api/views.py
+++ b/course/api/views.py
@@ -1,3 +1,5 @@
+from typing import Any, Dict, List, Union
+
 from rest_framework import filters, viewsets, status, mixins
 from rest_framework.decorators import action
 from rest_framework.exceptions import ParseError
@@ -12,6 +14,7 @@ from django.utils.translation import gettext_lazy as _
 
 from aplus.api import api_reverse
 from edit_course.operations.configure import configure_from_url
+from exercise.cache.content import ModuleEntry, ExerciseEntry
 from lib.api.constants import REGEX_INT, REGEX_INT_ME
 from lib.api.filters import FieldValuesFilter
 from lib.api.mixins import ListSerializerMixin, MeUserMixin
@@ -201,21 +204,25 @@ class CourseExercisesViewSet(NestedViewSetMixin,
     lookup_value_regex = REGEX_INT
     parent_lookup_map = {'course_id': 'course_instance.id'}
 
-    def __recurse_exercises(self, module, exercises):
-        for child in filter(lambda ex: ex['type'] == 'exercise', module['children']):
-            if child['submittable']:
+    def __recurse_exercises(
+            self,
+            module: Union[ModuleEntry, ExerciseEntry],
+            exercises: List[Dict[str, Any]],
+            ) -> List[Dict[str, Any]]:
+        for child in module.children:
+            if child.submittable:
                 exercise_dictionary = {
-                    'id': child['id'],
+                    'id': child.id,
                     'url': build_aplus_url(
-                        api_reverse('exercise-detail', kwargs={'exercise_id': child['id']}),
+                        api_reverse('exercise-detail', kwargs={'exercise_id': child.id}),
                         True,
                     ),
-                    'html_url': build_aplus_url(child['link'], True),
-                    'display_name': child['name'],
-                    'max_points': child['max_points'],
-                    'max_submissions': child['max_submissions'],
-                    'hierarchical_name': child['hierarchical_name'],
-                    'difficulty': child['difficulty'],
+                    'html_url': build_aplus_url(child.link, True),
+                    'display_name': child.name,
+                    'max_points': child.max_points,
+                    'max_submissions': child.max_submissions,
+                    'hierarchical_name': child.hierarchical_name,
+                    'difficulty': child.difficulty,
                 }
                 exercises.append(exercise_dictionary)
 
@@ -225,36 +232,36 @@ class CourseExercisesViewSet(NestedViewSetMixin,
 
         return exercises
 
-    def __module_to_dict(self, module, **kwargs):
-        kwargs['exercisemodule_id'] = module['id']
+    def __module_to_dict(self, module: ModuleEntry, **kwargs) -> Dict[str, Any]:
+        kwargs['exercisemodule_id'] = module.id
         module_dictionary = {
-            'id': module['id'],
+            'id': module.id,
             'url': build_aplus_url(api_reverse("course-exercises-detail", kwargs=kwargs), True),
-            'html_url': build_aplus_url(module['link'], True),
-            'display_name': module['name'],
+            'html_url': build_aplus_url(module.link, True),
+            'display_name': module.name,
             'is_open': CourseModule.check_is_open(
-                module['reading_opening_time'],
-                module['opening_time'],
-                module['closing_time'],
+                module.reading_opening_time,
+                module.opening_time,
+                module.closing_time,
             ),
         }
         module_dictionary['exercises'] = self.__recurse_exercises(module, [])
         return module_dictionary
 
-    def list(self, request, *args, **kwargs):
+    def list(self, request, *args, **kwargs) -> Response:
         modules = []
-        for module in self.content.data['modules']:
-            if module['status'] != CourseModule.STATUS.HIDDEN:
+        for module in self.content.data.modules:
+            if module.status != CourseModule.STATUS.HIDDEN:
                 modules.append(self.__module_to_dict(module, **kwargs))
         return Response({"count": len(modules), "next": None, "previous": None, 'results': modules})
 
-    def retrieve(self, request, *args, **kwargs):
+    def retrieve(self, request, *args, **kwargs) -> Response:
         # try to get the module list index
-        idx = self.content.data['module_index'].get(int(kwargs['exercisemodule_id']))
-        if idx is None:
+        entry = self.content.data.module_index.get(int(kwargs['exercisemodule_id']))
+        if entry is None:
             raise Http404()
 
-        return Response(self.__module_to_dict(self.content.data['modules'][idx[0]], **kwargs))
+        return Response(self.__module_to_dict(entry, **kwargs))
 
 
 class CourseExerciseTreeViewSet(CourseResourceMixin,
@@ -279,7 +286,7 @@ class CourseExerciseTreeViewSet(CourseResourceMixin,
 
     def list(self, request, *args, **kwargs):
         serializer = self.serializer_class(
-            self.content.data['modules'],
+            self.content.data.modules,
             many=True,
             context={ 'request': request }
         )