--- a/course/models.py
+++ b/course/models.py
@@ -547,8 +547,20 @@ def build_upload_dir(instance, filename):
     )
 
 
+class CourseInstanceProto(UrlMixin):
+    ABSOLUTE_URL_NAME = "course"
+    EDIT_URL_NAME = "course-edit"
+    url: str
+
+    def get_url_kwargs(self):
+        return {"instance_slug": self.url, **self.get_course_url_kwargs()}
+
+    def get_course_url_kwargs(self):
+        raise NotImplementedError(f"{self.__class__} must implement get_course_url_kwargs")
+
+
 @register_jwt_accessible_class("instance")
-class CourseInstance(UrlMixin, models.Model):
+class CourseInstance(CourseInstanceProto, models.Model):
     """
     CourseInstance class represent an instance of a course. A single course may have
     several instances either at the same time or during different years.
@@ -590,7 +602,7 @@ class CourseInstance(UrlMixin, models.Model):
         verbose_name=_('LABEL_INSTANCE_NAME'),
         max_length=255,
     )
-    url = models.CharField(
+    url: str = models.CharField( # type: ignore
         verbose_name=_('LABEL_URL'),
         max_length=255,
         blank=False,
@@ -1128,14 +1140,8 @@ class CourseInstance(UrlMixin, models.Model):
     def head_js_urls(self):
         return [url for url in self.head_urls.split() if ".js" in url]
 
-    ABSOLUTE_URL_NAME = "course"
-    EDIT_URL_NAME = "course-edit"
-
-    def get_url_kwargs(self):
-        # dict(foo=bar, **baz()) is not nice, but it's cleanest solution for this
-        # specific problem. For more read out stackoverflow answer about merging
-        # python dicts in single line: http://stackoverflow.com/a/26853961
-        return dict(instance_slug=self.url, **self.course.get_url_kwargs()) # pylint: disable=use-dict-literal
+    def get_course_url_kwargs(self) -> Dict[str, str]:
+        return self.course.get_url_kwargs()
 
     # A ConfigPart object is cached for each instance during an update. Cache is cleared
     # when the instance is deleted. The edit_course_operations.configure logic is responsible
@@ -1230,7 +1236,16 @@ class CourseModuleManager(models.Manager):
         return self.all()
 
 
-class CourseModule(UrlMixin, models.Model):
+class CourseModuleProto(UrlMixin):
+    ABSOLUTE_URL_NAME = "module"
+    course_instance: CourseInstanceProto
+    url: str
+
+    def get_url_kwargs(self):
+        return {"module_slug": self.url, **self.course_instance.get_url_kwargs()}
+
+
+class CourseModule(CourseModuleProto, models.Model):
     """
     CourseModule objects connect chapters and learning objects to logical sets
     of each other and course instances. They also contain information about the
@@ -1255,7 +1270,7 @@ class CourseModule(UrlMixin, models.Model):
         verbose_name=_('LABEL_NAME'),
         max_length=255,
     )
-    url = models.CharField(
+    url: str = models.CharField( # type: ignore
         verbose_name=_('LABEL_URL'),
         max_length=255,
         help_text=_('MODULE_URL_IDENTIFIER_HELPTEXT'),
@@ -1268,7 +1283,7 @@ class CourseModule(UrlMixin, models.Model):
         verbose_name=_('LABEL_INTRODUCTION'),
         blank=True,
     )
-    course_instance = models.ForeignKey(CourseInstance,
+    course_instance: CourseInstance = models.ForeignKey(CourseInstance, # type: ignore
         verbose_name=_('LABEL_COURSE_INSTANCE'),
         on_delete=models.CASCADE,
         related_name="course_modules",
@@ -1432,11 +1447,6 @@ class CourseModule(UrlMixin, models.Model):
     def number_of_submitters(self):
         return self.course_instance.students\
             .filter(submissions__exercise__course_module=self).distinct().count()
-
-    ABSOLUTE_URL_NAME = "module"
-
-    def get_url_kwargs(self):
-        return dict(module_slug=self.url, **self.course_instance.get_url_kwargs()) # pylint: disable=use-dict-literal
 
 
 class LearningObjectCategory(models.Model):