--- a/course/viewbase.py
+++ b/course/viewbase.py
@@ -39,6 +39,7 @@ class CourseInstanceBaseMixin:
     course_permission_classes = (
         CourseVisiblePermission,
     )
+    content: CachedContent
 
     def get_permissions(self):
         perms = super().get_permissions()