--- a/course/viewbase.py
+++ b/course/viewbase.py
@@ -67,6 +67,14 @@ class CourseInstanceBaseMixin:
         return CachedPoints(self.instance, self.request.user, self.is_course_staff)
 
     @property
+    def instance_max_group_size(self) -> int:
+        return self.content_bare.total().max_group_size
+
+    @property
+    def instance_min_group_size(self) -> int:
+        return self.content_bare.total().min_group_size
+
+    @property
     def course(self) -> Course:
         return self.instance.course