--- a/exercise/cache/exceptions.py
+++ b/exercise/cache/exceptions.py
@@ -1,0 +1,2 @@
+class NoSuchContent(Exception):
+    pass