… course/templates/course/_course_menu.html
… course/templates/course/course_base.html
… course/templates/course/groups.html
… diploma/grade.py
… diploma/models.py
… diploma/tests.py
… diploma/views.py
… edit_course/views.py
… exercise/api/csv/aggregate_points.py
… exercise/api/csv/aggregate_sheet.py
… exercise/api/csv/views.py
… exercise/api/custom_serializers.py
… exercise/api/full_serializers.py
… exercise/cache/basetypes.py
… exercise/cache/content.py
… exercise/cache/hierarchy.py
… exercise/cache/invalidate_util.py
… exercise/cache/points.py
… exercise/exercise_models.py
… exercise/exercise_summary.py
… exercise/exercisecollection_models.py
… exercise/management/commands/export.py
… exercise/reveal_states.py
… exercise/staff_views.py
… exercise/submission_models.py
… exercise/templates/exercise/_category_points.html
… exercise/templates/exercise/_children.html
… exercise/templates/exercise/_exercise_info.html
… exercise/templates/exercise/_points_badge.html
… exercise/templates/exercise/_submission_list.html