# Description

**What?**

- Refactors CachedContent/Points into separate objects: each module and exercise is stored separately. Allows loading points for individual exercises or loading points for the instance without loading each exercise and module points as well.
- Adds type hints to the cache entries.
- Replaces UserExerciseSummary with the cached SubmittableExerciseEntry.
- Transaction support for cache operations: cache changes are only committed if the transaction is.
- Support for loading multiple cache entries at once

BEHAVIOUR CHANGES:
- course front page now shows "Not submitted" on confirm_the_level exercises without submissions or confirmable points
  - matches the exercise pages
  - earlier it showed nothing
- points_badge submission status:
  - shows submission status as long as it hasn't been graded
  - earlier submission object wouldn't show status if confirm_the_level and status == Submission.STATUS.WAITING (in which case it showed "Submitted")
    - i.e. "In grading" vs "Submitted" for submissions that have confirm_the_level and status == Submission.STATUS.WAITING
- exercise page points badge now links to the best unofficial submission instead of first when all submissions are unofficial
  - seemed to be a bug in UserExerciseSummary
- exercise is 'passed' in cache only if best submission is READY but UserExerciseSummary checked that it is not UNOFFICIAL
  - ResultTable also only checks READY submissions when calculating the grade for an exercise
- missing_points defaulted to False when unrevealed for summary but True for cached (if points_to_pass > 0)
  - doesnt matter as the missing_points value isnt used when unrevealed
- a group is 'fixed' if there are any submissions to an exercise
  - previously only non-error submissions counted
- exercise boxes do not show unofficial submissions as part of the out of max submissions count
- course front page no longer shows unofficial points for exercises if only unofficial submissions exist
  - this is how it worked for exercise pages
- course front page only shows points for visible exercises
- the points of non-submittable exercises in the cache now depent on all descendants and not only the immediate children
 - e.g. a chapter's points used to only include the points of submittable exercises that were direct children. Now they also include the points of submittable exercises in children's children and their children etc.

**Why?**

- Points calculation code was duplicated in UserExerciseSummary (and still is elsewhere, see #888).
- There were differences in the points shown in UI on the front page and the exercise pages.
- CachedContent (one for each instance) was duplicated in each CachedPoints (one for each user for each instance)
- It wasn't very clear what is contained in the cache as there were no type hints, and the generation code was a bit confusing
- CachedContent and CachedPoints sometimes grew too large to be saved in the cache.
- Transaction support:
  - this was originally required for tests to pass but might not actually be needed anymore with the dependency checks
  - the cache isn't invalidated unneccessarily if the transaction is rolled back
  - newly generated cache entries might be based on invalid data and aren't saved if the transaction is rolled back
  - write to cache happens all at once at the end of the transaction: more performant
  - invalidation only happens at the end of the transaction which is were the changes actually happen to the database
    - it was possible before that the cache was invalidated due to database changes, then generated (by a different request) and the database changes were committed afterwards, leaving the cache in an invalid state.

**How?**

Introduces a new cache base class:
- Enables type hints for cache entries
- Each type hinted attribute is saved to cache (except if they are marked with NoCache, ClassVar or InitVar)
- Supports saving objects in the inheritance tree separately, allowing inheriting from other cache objects without duplicating the contents in the cache.
- Supports dependencies between cache entries, allowing an entry to be considered invalid if any of the dependencies are newer.
  - Makes invalidation handling much easier and reliable for complex dependencies

New ProxyManager class:
- Supports loading multiple cache entries at once, making loading _much_ faster when accessing multiple objects
- Tries to minimize the number of cache reads and writes for performance
- Enables separate cache entries to contain references to the same cache entry objects
  - Enables circular references and loading a cache entry once loads it for every object that references it

New DBDataManager class:
- Allows loading data from the database for multiple cache entries at once. Useful when the cache entries make use of the same database data.

Fixes a part of #888

# Testing

**What type of test did you run?**

- [ ] Accessibility test using the [WAVE](https://wave.webaim.org/extension/) extension.
- [X] Django unit tests.
- [X] Selenium tests.
- [ ] Other test. *(Add a description below)*
- [X] Manual testing.

Accessed most of the major pages while debugging,

**Did you test the changes in**

- [X] Chrome
- [X] Firefox
- [ ] This pull request cannot be tested in the browser.

# Translation

- [ ] Did you modify or add new strings in the user interface? ([Read about how to create translation](https://github.com/apluslms/a-plus/tree/master/doc#running-tests-and-updating-translations))

# Programming style

- [X] Did you follow our [style guides](https://apluslms.github.io/contribute/styleguides/)?
- [X] Did you use Python type hinting in all functions that you added or edited? ([type hints](https://docs.python.org/3/library/typing.html) for function parameters and return values)

# Have you updated the README or other relevant documentation?

- [ ] documents inside the doc directory.
- [ ] README.md.
- [ ] Aplus Manual.
- [ ] Other documentation (mention below which documentation).

# Is it Done?

- [ ] Reviewer has finished the code review
- [ ] After the review, the developer has made changes accordingly
- [ ] Customer/Teacher has accepted the implementation of the feature