@lainets 2023-08-18
A new issue to be added if this is merged: `exercise.viewbase.ExerciseMixin.get_summary_submissions` passes ? self.feedback_revealed` to `ExercisePoints` but I believe `self.is_course_staff` would be enough (wouldn't require DB access then). This needs to be investigated. This would also make it nearly identical to the `cached_points` property in the `get_cached_points` method: only difference is that `cached_points` can technically be for a non-submittable exercise. Would also need to investigate whether these two can be merged (is the exercise always submittable?)