@lainets 2023-08-18
## Performance changes

I'm a little worried about the large variances in the load times. Sometimes it seems to be very slow. I'm not sure if this only happens with the new changes but it does seem like that. It might be caused by the garbage collector activating during the load.

### Synopsis

- No apparent change to course configure time.
- Load time: Slower than before and very much dependent on the number of separate cache entries to be loaded (i.e. the number of LearningObjects on the course). This means the impact is seen more on large courses. CachedContent went from ~5ms to ~25ms on O1, and points went from 25-30ms to ~44ms. Sometimes it takes considerably more time (may be due to garbage collector).
- Generation time:
  - maybe a little faster for cached content
  - seemingly a little slower for cached points for few submissions
  - seemingly faster for cached points for a lot of submissions

### O1 benchmarks (on minus)

The decimal point numbers are in seconds. `4-8ms ~5ms` means that the value varied from 4 to 8 ms but (estimated) median took 5ms.

BEFORE (pymemcache):
content: 
- generate: 0.740134373
- load: 4-8ms ~5ms
Points with no submissions:
- generate: 0.15735532900000004 
- load: 21-27ms ~27ms
Points with few hundred submissions:
- generate: 0.609012841 
- load: 20-31ms ~30ms

BEFORE (pylibmc):
content:
- generate: - 0.563846949 0.9350565309999999
- load: 4-8ms ~7ms
Points with no submissions:
- generate: 0.14251463899999806 0.10672521400000001
- load: 0.14770852599999973 ~19ms
Points with few hundred submissions:
- generate: 0.8493996490000004 0.588477317 0.706424712
- load: ~24ms

AFTER (pymemcache):
content: 
- generate: 0.627027349 0.6389316300000001 
- load: 60-200ms ~60ms
Points with no submissions:
- generate: 0.222431614 0.25517966700000017 
- load: 80-200ms ~80ms
Points with few hundred submissions:
- generate: 0.46071420100000005 0.5961609270000001 
- load: 80-200ms ~86ms

AFTER (pylibmc):
content:
- generate: 0.66984569 0.46586513700000004
- load: ~25ms
Points with no submissions:
- generate: 0.1542314090000001 0.13666057200000004
- load: 40-270ms ~44ms
Points with few hundred submissions:
- generate: 0.5559999369999999 0.5619267869999999
- load: 40-270ms ~44ms