@Mankro 2023-08-31
I forgot to add in the previous message that it is great that you have written long git commit messages. They help others to understand all this.