@lainets 2023-09-01
Some basic benchmarks of fully loading (including exercises) a particular chapter page with multiple exercises on O1 on my pc:
- 2,25-2,6s before cache changes and without the optimization at the end of this PR
- 3,0-3,2s after cache changes and without optimization (this value might actually be old but it should be a little higher than the above)
- 2,00-2,2s after cache changes and with optimizations
- 0,95-1,1s after cache changes and with optimizations and loading the exercises in parallel instead of sequentially